refactor: put dynamic acl to a feature
